Bring up-to-date, directly call Viz from extension.ts, bump 0.1.0
This simplifies the setup, I assume old versions of vscode still
required a lot of workarounds.

Move to new Viz library which is a real module, so we can directly
call it from `extension.ts` instead of the workaround. This speeds up
things considerably, because we don’t have to initialize the wasm
library on each call, instead do it once on extension init.

The `<pre style="all:unset">` trick is the same as the one used by
the preview code generated by the mermaid extension.

I removed the test stubs.
I moved to npm lockfiles and removed yarn.

This bumps the extension version to 0.1.0.